requirements.txt
# Macro Dashboard — Private, local-first, newsletter-ready charts
pandas>=2.0.0
numpy>=1.24.0
fredapi>=0.5.0
yfinance>=0.2.0
plotly>=5.18.0
streamlit>=1.28.0
python-dotenv>=1.0.0
kaleido>=0.2.0
